Begin refactor of candle-metal-kernels
