+++ candle-metal-kernels/src/metal/device.rs
@@ -1,0 +1,105 @@
+use crate::{
+    Buffer, ComputePipeline, ConstantValues, Function, Library, MTLResourceOptions,
+    MetalKernelError,
+};
+use objc2::{rc::Retained, runtime::ProtocolObject};
+use objc2_foundation::{NSRange, NSString};
+use objc2_metal::{
+    MTLBlitCommandEncoder, MTLBuffer, MTLCommandBuffer, MTLCommandBufferStatus, MTLCommandQueue,
+    MTLCompileOptions, MTLComputeCommandEncoder, MTLComputePipelineState, MTLCounterSet,
+    MTLCreateSystemDefaultDevice, MTLDataType, MTLDevice, MTLFunction, MTLFunctionConstantValues,
+    MTLLibrary, MTLResource, MTLResourceUsage, MTLSize,
+};
+use std::{collections::HashMap, ffi::c_void, ptr, sync::Arc};
+
+// Use Retained when appropriate. Gives us a more elegant way of handling memory (peaks) than autoreleasepool.
+// https://docs.rs/objc2/latest/objc2/rc/struct.Retained.html
+pub type CommandQueue = Retained<ProtocolObject<dyn MTLCommandQueue>>;
+pub type CounterSet = Retained<ProtocolObject<dyn MTLCounterSet>>;
+
+#[derive(Clone, Debug)]
+pub struct Device {
+    raw: Retained<ProtocolObject<dyn MTLDevice>>,
+}
+unsafe impl Send for Device {}
+unsafe impl Sync for Device {}
+
+impl Device {
+    pub fn as_ref(&self) -> &ProtocolObject<dyn MTLDevice> {
+        &*self.raw
+    }
+
+    pub fn registry_id(&self) -> u64 {
+        self.as_ref().registryID()
+    }
+
+    pub fn all() -> Vec<Self> {
+        MTLCreateSystemDefaultDevice()
+            .into_iter()
+            .map(|raw| Device { raw })
+            .collect()
+    }
+
+    pub fn system_default() -> Option<Self> {
+        MTLCreateSystemDefaultDevice().map(|raw| Device { raw })
+    }
+
+    pub fn new_buffer(
+        &self,
+        length: usize,
+        options: MTLResourceOptions,
+    ) -> Result<Buffer, MetalKernelError> {
+        self.as_ref()
+            .newBufferWithLength_options(length, options)
+            .map(Buffer::new)
+            .ok_or(MetalKernelError::FailedToCreateResource(
+                "Buffer".to_string(),
+            ))
+    }
+
+    pub fn new_buffer_with_data(
+        &self,
+        pointer: *const c_void,
+        length: usize,
+        options: MTLResourceOptions,
+    ) -> Result<Buffer, MetalKernelError> {
+        let pointer = ptr::NonNull::new(pointer as *mut c_void).unwrap();
+        unsafe {
+            self.as_ref()
+                .newBufferWithBytes_length_options(pointer, length, options)
+                .map(Buffer::new)
+                .ok_or(MetalKernelError::FailedToCreateResource(
+                    "Buffer".to_string(),
+                ))
+        }
+    }
+
+    pub fn new_library_with_source(
+        &self,
+        source: &str,
+        options: Option<&MTLCompileOptions>,
+    ) -> Result<Library, MetalKernelError> {
+        let raw = self
+            .as_ref()
+            .newLibraryWithSource_options_error(&NSString::from_str(source), options)
+            .unwrap();
+
+        Ok(Library::new(raw))
+    }
+
+    pub fn new_compute_pipeline_state_with_function(
+        &self,
+        function: &Function,
+    ) -> Result<ComputePipeline, MetalKernelError> {
+        let raw = self
+            .as_ref()
+            .newComputePipelineStateWithFunction_error(function.as_ref())
+            .unwrap();
+        Ok(ComputePipeline::new(raw))
+    }
+
+    pub fn new_command_queue(&self) -> Result<CommandQueue, MetalKernelError> {
+        let raw = self.as_ref().newCommandQueue().unwrap();
+        Ok(raw)
+    }
+}
