+++ candle-metal-kernels/src/metal/mod.rs
@@ -1,0 +1,4 @@
+pub mod device;
+pub mod metal_utils;
+pub use device::*;
+pub use metal_utils::*;
